src/user/auth_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.6
email-validator==2.1.0
redis==5.0.1
aio-pika==9.3.0
alembic==1.13.1